+++ R/utils-pdftk.R
@@ -1,8 +1,11 @@
 # SPDX-License-Identifier: MIT
 
 get_pdftk_metadata <- function(filename) {
+    f <- tempfile(fileext = ".txt")
+    on.exit(unlink(f))
     cmd <- pdftk()
     filename <- shQuote(normalizePath(filename, mustWork = TRUE))
-    args <- c(filename, "dump_data_utf8")
-    xmpdf_system2(cmd, args)
+    args <- c(filename, "dump_data_utf8", "output", f)
+    results <- xmpdf_system2(cmd, args)
+    brio::read_lines(f)
 }