+++ R/zzz.R
@@ -3,4 +3,5 @@
 #' @importFrom rlang abort inform warn %||%
 #' @importFrom utils hasName
 #' @importFrom R6 R6Class
+#' @importFrom stringi stri_join
 NULL